+++ test/dragonfruit.service.ts
@@ -1,6 +1,6 @@
 import { ModuleRef } from '@nestjs/core';
 
-import { DragonFruit } from './entities/DragonXruit';
+import { DragonFruit } from './entities/DragonFruit';
 import { MyConfigService } from './myconfig.service';
 import { Injectable, BadRequestException } from '@nestjs/common';
 import { CrudService } from '@eicrud/core/crud/crud.service';
